% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLA_pop_freq.R
\docType{data}
\name{freq}
\alias{freq}
\title{HLA allele frequency in the Human Population}
\format{
An object of class "data.frame".
}
\usage{
scrHLAmatrix::freq
}
\description{
Data from the Allele Frequency Net Database (AFND) of gene frequencies in different worldwide populations.
The data has been downloaded from the AFND website (http://www.allelefrequencies.net/hla.asp) for classical and non classical HLA alleles; and is up to date as of 13-JUL-2023.
}
\examples{
hla_pop_freq <- scrHLAmatrix::freq
}
\keyword{datasets}
