% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load_counts.R
\name{HLA_load}
\alias{HLA_load}
\title{Loading the count file(s) as an output of scrHLAtag, including columns for CB, UMI, and HLA alleles (\url{https://github.com/furlan-lab/scrHLAtag}).}
\usage{
HLA_load(
  directories,
  dir_names = NULL,
  seu,
  scrHLAtag_outs = c("molecule_info_mRNA.txt.gz", "molecule_info_gene.txt.gz"),
  colnames = NULL
)
}
\arguments{
\item{directories}{a character list of one (or more) full-length directory(ies) where the the scrHLAtag count file(s) have been organized.}

\item{dir_names}{a character list of one (or more) sample name(s) that \emph{strictly correspond(s)} to the directory(ies) where the the scrHLAtag count file(s) have been organized. Can remain \code{NULL} only if the \code{directories} are already named vectors. For 2 or more samples/directories, try to get the names as closely as possible to the sample ID names spposedly present in the corresponding Seurat object (usually concatenated to Seurat barcodes when experiments are merged).}

\item{seu}{is the Seurat object associated with the scrHLAtag count file (\url{https://satijalab.org/seurat/index.html}).}

\item{scrHLAtag_outs}{a character list of the proper scrHLAtag count file(s) to be loaded. For example, for each sample, the desired count files should be the \code{"molecule_info_mRNA.txt.gz"} and the \code{"molecule_info_gene.txt.gz"} files.}

\item{colnames}{\code{NULL} or a character list of the colnames of the scrHLAtag count file(s), if they are different from default output (scrHLAtag, version 0.1.5).}
}
\value{
a list of data.frames/data.tables
}
\description{
Loading the count file(s) as an output of scrHLAtag, including columns for CB, UMI, and HLA alleles (\url{https://github.com/furlan-lab/scrHLAtag}).
}
\examples{
dirs_path <- "path/to/scrHLAtag/out/files"
dirs<-list.dirs(path=dirs_path, full.names = T, recursive = F)
dirs<- lapply(dirs, list.dirs, recursive = F) \%>\% unlist
dirs<- lapply(dirs, dir, pattern = "unguided_hla_align_corrected", recursive = F, full.names = T) \%>\% unlist
dirnames <- c("AML_101_BM", "AML_101_34", "TN_BM", "TN_34") # this is how the samples were organized in the directories
## Load the counts files
cts <- HLA_load(directories = dirs, dir_names = dirnames, seu = your_Seurat_obj)
}
