% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrHLAmatrix.R
\name{Top_HLA_list}
\alias{Top_HLA_list}
\title{Extracting the top HLA alleles from the scrHLAtag count files}
\usage{
Top_HLA_list(
  cts_list,
  frac = 0.65,
  min_alleles_keep = 5,
  min_reads_per_gene = 200,
  insert_pop_most_freq = TRUE,
  use_gene_align_ABC = FALSE
)
}
\arguments{
\item{cts_list}{is a list of 2 scrHLAtag count files including columns for CB, UMI, and HLA alleles. The first contains genomic molecular info (gene) and the second contains mRNA molecular info (https://github.com/furlan-lab/scrHLAtag).}

\item{frac}{is the fraction (0 to 1) of total reads for a particular HLA gene, which incorporates the highest ranking alleles of that gene in terms of number of reads; default at 0.65 .}

\item{min_alleles_keep}{is a numeric representing the minimum number of highest ranking alleles to keep despite filtering by fraction 'frac'; default is 5.}

\item{min_reads_per_gene}{is a numeric representing minimum number of total reads per HLA gene (including all its alleles) below which the gene is filtered out; default is 200.}

\item{insert_pop_most_freq}{is a logical, whether to to include the HLA allele with the highest frequency in the human population despite low reads in the count files; default is TRUE.}

\item{use_gene_align_ABC}{is a logical, whether to use the count file from the genomic alignment (rather than the mRNA alignment) to count reads for the HLA-A, -B, and -C genes, as it was observed in some cases that this has better accuracy in predicting genotype versus mRNA alignments (not the case for Class-II and other HLA genes); default is FALSE.}
}
\value{
a Vector of the top HLA alleles in the count files (in terms of reads).
}
\description{
Extracting the top HLA alleles from the scrHLAtag count files
}
\examples{
Top_HLA_list(cts_list = cts, frac = 0.8, min_alleles_keep = 2, use_gene_align_ABC = TRUE)
}
