% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrHLAmatrix.R
\name{Top_HLA_plot}
\alias{Top_HLA_plot}
\title{Plotting the top HLA alleles from the scrHLAtag count files}
\usage{
Top_HLA_plot(
  cts_list,
  top_hla = 10,
  min_reads_per_gene = 200,
  use_gene_align_ABC = FALSE
)
}
\arguments{
\item{cts_list}{is a list of 2 scrHLAtag count files including columns for CB, UMI, and HLA alleles. The first contains genomic molecular info (gene) and the second contains mRNA molecular info (https://github.com/furlan-lab/scrHLAtag).}

\item{top_hla}{is a numeric, representing the number of top HLA alleles (i.e. with the highest number of reads) per HLA gene to display in the plot; default is 10.}

\item{min_reads_per_gene}{is a numeric representing minimum number of total reads per HLA gene (including all its alleles) below which the gene is filtered out; default is 200.}

\item{use_gene_align_ABC}{is a logical, whether to use the count file from the genomic alignment (rather than the mRNA alignment) to count reads for the HLA-A, -B, and -C genes, as it was observed in some cases that this has better accuracy in predicting genotype versus mRNA alignments (not the case for Class-II and other HLA genes); default is FALSE.}
}
\value{
a Vector of the top HLA alleles in the count files (in terms of reads).
}
\description{
Plotting the top HLA alleles from the scrHLAtag count files
}
\examples{
Top_HLA_plot(cts_list = cts, use_gene_align_ABC = TRUE)
}
