% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Top_HLA_alleles.R
\name{Top_HLA_plot_bulk}
\alias{Top_HLA_plot_bulk}
\title{Plotting the top HLA alleles extracted in a Pseudo-Bulk approach from the scrHLAtag count files}
\usage{
Top_HLA_plot_bulk(
  reads_1,
  reads_2 = NULL,
  cluster_index = NULL,
  top_hla = 10,
  min_reads_per_gene = 20,
  use_alt_align_ABC = FALSE,
  color_pal = NULL
)
}
\arguments{
\item{reads_1}{is the primary scrHLAtag count file (1 of 2 files containing either the mRNA molecular info or the genomic (gene) molecular info). It includes columns for CB, UMI, and HLA alleles (\url{https://github.com/furlan-lab/scrHLAtag}).}

\item{reads_2}{is the secondary scrHLAtag count file (the alternative file vs. the one designated in \code{reads_1} argument). It includes columns for CB, UMI, and HLA alleles (\url{https://github.com/furlan-lab/scrHLAtag}). Default is \code{NULL}, in which case it will not be able to count alternative aligment and argument \code{use_alt_align_ABC} will become irrelevant.}

\item{cluster_index}{a numeric or integer, representing the index of the HLA cluster present in the counts data (as previously analyzed by \code{HLA_clusters()} and mapped back to the data by \code{map_HLA_clusters()}). It allows to subset the visualization based on the selected HLA cluster. \code{NULL} is the default in which case the counts data will not be subsetted.}

\item{top_hla}{numeric, representing the number of top HLA alleles (i.e. with the highest number of reads) per HLA gene to display in the plot; default is \code{10}.}

\item{min_reads_per_gene}{numeric representing minimum number of total reads per HLA gene (including all its alleles) below which the gene is filtered out; default is \code{20}.}

\item{use_alt_align_ABC}{logical, whether to use the count file from the alternative alignment (rather than the primary alignment) to count reads for the HLA-A, -B, and -C genes. It was observed in some cases that using genomic alignments has better accuracy in predicting genotype versus mRNA alignments (not the case for Class-II and other HLA genes); default is \code{FALSE}.}

\item{color_pal}{is a character list of colors to visualize HLA polulation frequencies when available. When \code{color_pal} is not provided (\code{NULL}), it defaults to \code{viridis::viridis(n = 10, option = "C")}.}
}
\value{
a Vector of the top HLA alleles in the count files (in terms of reads).
}
\description{
Plotting the top HLA alleles extracted in a Pseudo-Bulk approach from the scrHLAtag count files
}
\examples{
dirs_path <- "path/to/scrHLAtag/out/files"
dirs<-list.dirs(path=dirs_path, full.names = T, recursive = F)
dirs<- lapply(dirs, list.dirs, recursive = F) \%>\% unlist
dirs<- lapply(dirs, dir, pattern = "unguided_hla_align_corrected", recursive = F, full.names = T) \%>\% unlist
dirnames <- c("AML_101_BM", "AML_101_34", "TN_BM", "TN_34") # this is how the samples were organized in the directories
## Load the counts files
cts <- HLA_load(directories = dirs, dir_names = dirnames, seu = your_Seurat_obj)
## Process those count files
Top_HLA_plot_bulk(reads_1 = cts[["mRNA"]], reads_2 = cts[["gene"]], use_alt_align_ABC = TRUE)
}
