% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Top_HLA_alleles.R
\name{Top_HLA_plot_byCB}
\alias{Top_HLA_plot_byCB}
\title{Getting the top two alleles (for each HLA gene) with the most counts per Cell Barcode}
\usage{
Top_HLA_plot_byCB(
  reads,
  cell_data_obj = NULL,
  CB_rev_com = FALSE,
  hla_with_counts_above = 5,
  CBs_with_counts_above = 15,
  match_CB_with_obj = TRUE,
  cluster_index = NULL,
  top_hla = 10,
  field_resolution = 3,
  QC_mm2 = TRUE,
  s1_percent_pass_score = 80,
  AS_percent_pass_score = 80,
  NM_thresh = 15,
  de_thresh = 0.01,
  default_theme = TRUE,
  return_genotype_data = FALSE,
  parallelize = TRUE
)
}
\arguments{
\item{reads}{is the scrHLAtag count file including columns for CB, UMI, and HLA alleles (\url{https://github.com/furlan-lab/scrHLAtag}).}

\item{cell_data_obj}{is a cell dataset object associated with the scrHLAtag count file; entered if matching CBs in count file with ones in the object is desired. Currently the function is compatible with Seurat (\url{https://satijalab.org/seurat/index.html}).}

\item{CB_rev_com}{logical, called \code{TRUE} if the need to obtain the reverse complement of Cell Barcodes (CBs) is desired; default is \code{FALSE}.}

\item{hla_with_counts_above}{the number of total reads accross CBs at or above which an HLA allele is retained in the matrix.}

\item{CBs_with_counts_above}{the number of total reads accross HLA alleles at or above which a CB is retained in the matrix. Note: at present, the function will make sure that number of CBs is equal or more than available HLA alleles in the matrix.}

\item{match_CB_with_obj}{is a logical, called TRUE if filtering CBs in the scrHLAtag count file with matching ones in the cell dataset object is desired.}

\item{cluster_index}{a numeric or integer, representing the index of the HLA cluster present in the counts data (as previously analyzed by \code{HLA_clusters()} and mapped back to the data by \code{map_HLA_clusters()}). It allows to subset the visualization based on the selected HLA cluster. \code{NULL} is the default in which case the counts data will not be subsetted.}

\item{top_hla}{the number of top HLA alleles (i.e. with the highest number of reads) per HLA gene to display in the plot; default is \code{10}.}

\item{field_resolution}{integer fron \code{1} to \code{3}, to select the HLA nomenclature level of Field resolution, where \code{1}, \code{2}, or \code{3} will take into consideration the first, the first two, or the first three field(s) of HLA designation; default is \code{3}.}

\item{QC_mm2}{logical, called \code{TRUE} if removing low quality reads based on minimap2 tags is desired.}

\item{s1_percent_pass_score}{percentage, \code{0} to \code{100}, cuttoff from the maximum score (best quality) of the minimap2 's1' tag, which a read needs to acheive to pass as acceptable; default at \code{80} and becomes less inclusive if value increases.}

\item{AS_percent_pass_score}{percentage, \code{0} to \code{100}, cuttoff from the maximum score (best quality) of the minimap2 'AS' tag, which a read needs to acheive to pass as acceptable; default at \code{80} and becomes less inclusive if value increases.}

\item{NM_thresh}{number of mismatches and gaps in the minimap2 alignment at or below which the quality of the read is acceptable; default is \code{15}.}

\item{de_thresh}{gap-compressed per-base sequence divergence in the minimap2 alignment at or below which the quality of the read is acceptable; the number is between \code{0} and \code{1}, and default is \code{0.01}.}

\item{default_theme}{logical, called \code{TRUE} by default to output a 'plot_grid' of all HLA alleles in the count file in a single figure; if \code{FALSE} the output is a list of plots without a specified theme.}

\item{return_genotype_data}{logical, called \code{TRUE} if the HLA genotypes of each CB, ordered in a dataframe, is desired in addition to the plots; \code{FALSE} by default.}

\item{parallelize}{logical, called \code{TRUE} if using parallel processing (multi-threading) is desired; default is \code{TRUE}.}
}
\value{
ggplot of HLA genotypes per Cell Barcodes and the number of times they occur
}
\description{
Getting the top two alleles (for each HLA gene) with the most counts per Cell Barcode
}
\examples{
dirs_path <- "path/to/scrHLAtag/out/files"
dirs<-list.dirs(path=dirs_path, full.names = T, recursive = F)
dirs<- lapply(dirs, list.dirs, recursive = F) \%>\% unlist
dirs<- lapply(dirs, dir, pattern = "unguided_hla_align_corrected", recursive = F, full.names = T) \%>\% unlist
dirnames <- c("AML_101_BM", "AML_101_34", "TN_BM", "TN_34") # this is how the samples were organized in the directories
## Load the counts files
cts <- HLA_load(directories = dirs, dir_names = dirnames, cell_data_obj = your_cell_dataset_obj)
## Process those count files
top_per_cb_plot <- Top_HLA_plot_byCB(reads = cts[["mRNA"]], cell_data_obj = your_cell_dataset_obj, hla_with_counts_above = 5, CBs_with_counts_above = 35)
top_per_cb_plot \%>\% gridExtra::grid.arrange()
# 
# Note: if for a particular HLA, the alleles with the most counts are in a tie 
# between 3 or more alleles in a particular Cell Barcode, we cannot know which 
# ones are the top two alleles, so that CB is not counted. This is similar if  
# there were no counts for that allele (all zeros).
}
