% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrHLAmatrix.R
\name{HLA_Matrix}
\alias{HLA_Matrix}
\title{Deduping and Correcting scrHLAtag counts and Creating Seurat-compatible Matrices}
\usage{
HLA_Matrix(cts, seu, rp = character(), dn = character(), Ct = 0)
}
\arguments{
\item{cts}{is the scrHLAtag count file including columns for CB, UMI, and HLA alleles (https://github.com/furlan-lab/scrHLAtag)}

\item{seu}{is the Seurat object associated with the scrHLAtag count file (https://satijalab.org/seurat/index.html)}

\item{rp}{is a character list of recipient-specific HLA alleles if known}

\item{dn}{is a character list of donor-specific HLA alleles if known}

\item{Ct}{is the count threshold for the PCR copies of UMIs to retain, default at 0}
}
\value{
an Assay type matrix
}
\description{
Deduping and Correcting scrHLAtag counts and Creating Seurat-compatible Matrices
}
\examples{
HLA_Matrix(cts = your_counts_df, seu = your_Seurat_obj, rp = c("A*24:02:01", "DRB1*04:01:01", "DRB4*01:03:02"), dn = c("A*33:03:01", "B*42:01:01"))
}
