% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlecell_genotyping.R
\name{Top_HLA_plot_byCB}
\alias{Top_HLA_plot_byCB}
\title{Getting the top two alleles (for each HLA gene) with the most counts per Cell Barcode}
\usage{
Top_HLA_plot_byCB(
  reads,
  seu = NULL,
  CB_rev_com = FALSE,
  hla_with_counts_above = 5,
  CBs_with_counts_above = 15,
  match_CB_with_seu = TRUE,
  cluster_index = NULL,
  top_hla = 10,
  field_resolution = 3,
  QC_mm2 = TRUE,
  s1_belowmax = 0.8,
  AS_belowmax = 0.8,
  NM_thresh = 15,
  de_thresh = 0.01,
  default_theme = TRUE,
  return_genotype_data = FALSE,
  parallelize = TRUE
)
}
\arguments{
\item{reads}{is the scrHLAtag count file including columns for CB, UMI, and HLA alleles (https://github.com/furlan-lab/scrHLAtag).}

\item{seu}{is the Seurat object associated with the scrHLAtag count file (https://satijalab.org/seurat/index.html), and entered here if matching CBs in count file with Seurat colnames is desired.}

\item{CB_rev_com}{is a logical, called TRUE if the need to obtained the reverse complement of Cell Barcodes (CBs) is desired; default is FALSE.}

\item{hla_with_counts_above}{is the number of total reads accross CBs at or above which an HLA allele is retained in the matrix.}

\item{CBs_with_counts_above}{is the number of total reads accross HLA alleles at or above which a CB is retained in the matrix. Note: at present, the function will make sure that number of CBs is equal or more than available HLA alleles in the matrix.}

\item{match_CB_with_seu}{is a logical, called TRUE if filtering CBs in the scrHLAtag count file with matching ones in the Seurat object is desired.}

\item{cluster_index}{is a numeric, representing the index of the HLA cluster present in the counts data (as previously analyzed by 'HLA_clusters()' and mapped back to the data by 'map_HLA_clusters()'). It allows to subset the visualization based on the selected HLA cluster. 'NULL' is the default in which case the counts data will not be subsetted.}

\item{top_hla}{is a numeric, representing the number of top HLA alleles (i.e. with the highest number of reads) per HLA gene to display in the plot; default is 10.}

\item{field_resolution}{is a numeric, to select the HLA nomenclature level of Field resolution, where 1, 2, or 3 will take into consideration the first, the first two, or the first three field(s) of HLA designation; default is 3.}

\item{QC_mm2}{is a logical, called TRUE if removing low quality reads based on minimap2 tags is desired.}

\item{s1_belowmax}{is a proportion (0 to 1) of the maximum value (best quality) of the minimap2 's1' tag above which the quality of the read is acceptable; default at 0.75 of the max s1 score.}

\item{AS_belowmax}{is a proportion (0 to 1) of the maximum value (best quality) of the minimap2 'AS' tag above which the quality of the read is acceptable; default at 0.85 of the max AS score.}

\item{NM_thresh}{is the number of mismatches and gaps in the minimap2 alignment at or below which the quality of the read is acceptable; default is 15.}

\item{de_thresh}{is the gap-compressed per-base sequence divergence in the minimap2 alignment at or below which the quality of the read is acceptable; the number is between 0 and 1, and default is 0.015.}

\item{default_theme}{is a logical, called TRUE by default to output a 'plot_grid' of all HLA alleles in the count file in a single figure; if FALSE the output is a list of plots without a specified theme.}

\item{return_genotype_data}{is a logical, called TRUE if the HLA genotypes of each CB, ordered in a dataframe, is desired in addition to the plots; FALSE by default.}

\item{parallelize}{is a logical, called TRUE if using parallel processing (multi-threading) is desired; default is TRUE.}
}
\value{
ggplot of HLA genotypes per Cell Barcodes and the number of times they occur
}
\description{
Getting the top two alleles (for each HLA gene) with the most counts per Cell Barcode
}
\examples{
samples <- c("AML_101_BM", "AML_101_34")
mol_info <- c("molecule_info_gene.txt.gz", "molecule_info_mRNA.txt.gz")
for (i in 1:length(mol_info)){
  dl<-lapply(samples, function(sample){
    d<-read.table(file.path("path/to/scrHLAtag/out/files", sample,
                            mol_info[i]), header = F, sep=" ", fill = T) 
    d$V1<-paste0(sample, "_", d$V1, "-1")
    colnames(d)<-c("name","CB", "nb", "UMI", "gene", "query_len","start", "mapq", "cigar", "NM", "AS", "s1", "de", "seq")
    d$samp <- sample
    d
  })
  ctsu<-do.call(rbind,dl)
  rm(dl)
  cts[[str_sub(strsplit(mol_info[i], "\\\\.")[[1]][1], start= -4)]] <- ctsu
  rm(ctsu)
}
Top_HLA_plot_byCB(reads = cts[["mRNA"]], seu = your_Seurat_Obj, hla_with_counts_above = 5, CBs_with_counts_above = 35)
# 
# Note: if for a particular HLA, the alleles with the most counts are in a tie 
# between 3 or more alleles in a particular Cell Barcode, we cannot know which 
# ones are the top two alleles, so that CB is not counted. This is similar if  
# there were no counts for that allele (all zeros).
}
